Add/drop PK

+++ go/vt/sqlparser/ast.go
@@ -2288,18 +2288,24 @@ type IndexSpec struct {
 	Columns []*IndexColumn
 	// Options contains the index options when creating an index
 	Options []*IndexOption
+	// Primary indicates a PRIMARY KEY create or delete operation
+	Primary bool
 }
 
 func (idx *IndexSpec) Format(buf *TrackedBuffer) {
 	switch idx.Action {
 	case "create", "CREATE":
 		buf.Myprintf("add ")
-		if idx.Type != "" {
-			buf.Myprintf("%s ", idx.Type)
-		}
-		buf.Myprintf("index %s ", idx.ToName.val)
-		if idx.Using.val != "" {
-			buf.Myprintf("using %s ", idx.Using.val)
+		if idx.Primary {
+			buf.Myprintf("primary key ")
+        } else {
+			if idx.Type != "" {
+				buf.Myprintf("%s ", idx.Type)
+			}
+			buf.Myprintf("index %s ", idx.ToName.val)
+			if idx.Using.val != "" {
+				buf.Myprintf("using %s ", idx.Using.val)
+			}
 		}
 		buf.Myprintf("(")
 		for i, col := range idx.Columns {
@@ -2325,7 +2331,11 @@ func (idx *IndexSpec) Format(buf *TrackedBuffer) {
 			}
 		}
 	case "drop", "DROP":
-		buf.Myprintf("drop index %s", idx.ToName.val)
+		if idx.Primary {
+			buf.Myprintf("drop primary key")
+		} else {
+			buf.Myprintf("drop index %s", idx.ToName.val)
+		}
 	case "rename", "RENAME":
 		buf.Myprintf("rename index %s to %s", idx.FromName.val, idx.ToName.val)
 	}
